refactor(cybergym): nest the Harbor adapter under integrations/harbor
CyberGym is a Harbor compose adapter, not a second harness. Keep it next to HarborRuntime instead of a sibling integrations/ package.
