+++ .pre-commit-config.yaml
@@ -4,6 +4,6 @@ repos:
     hooks:
       - id: ruff
         args: ["--fix", "--show-fixes", "--output-format=full"]
-        exclude: ^.*\.(ipynb)$|^verl/.*$|^rllm/integrations/cybergym/harbor_adapter/.*$
+        exclude: ^.*\.(ipynb)$|^verl/.*$|^rllm/integrations/harbor/cybergym/harbor_adapter/.*$
       - id: ruff-format
-        exclude: ^.*\.(ipynb)$|^verl/.*$|^rllm/integrations/cybergym/harbor_adapter/.*$
+        exclude: ^.*\.(ipynb)$|^verl/.*$|^rllm/integrations/harbor/cybergym/harbor_adapter/.*$
