+++ rllm/integrations/harbor/cybergym/harbor_adapter/template/environment/docker-compose.yaml
