+++ rllm/integrations/harbor/cybergym/harbor_adapter/template/environment/restrict-network.sh
