+++ rllm/integrations/harbor/cybergym/harbor_adapter/template/environment/task-server/Dockerfile.oss_fuzz
