+++ rllm/integrations/harbor/cybergym/harbor_adapter/template/solution/solve.sh
